// src/lib/graphql/types.gen.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  DateTime: string
}
// ...
export type Comment = {
  __typename?: 'Comment'
  author: User
  body: Scalars['String']
  createdAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
  deletedBy?: Maybe<Scalars['Int']>
  id: Scalars['Int']
  old_id?: Maybe<Scalars['String']>
  old_thread?: Maybe<Scalars['String']>
  ratings?: Maybe<Array<Maybe<CommentRating>>>
  replyTo?: Maybe<Scalars['Int']>
  shout: Scalars['Int']
  updatedAt?: Maybe<Scalars['DateTime']>
  updatedBy?: Maybe<Scalars['Int']>
  views?: Maybe<Scalars['Int']>
}

export type CommentRating = {
  __typename?: 'CommentRating'
  comment_id: Scalars['Int']
  createdAt: Scalars['DateTime']
  createdBy: Scalars['String']
  id: Scalars['Int']
  value: Scalars['Int']
}
// ...
export type Rating = {
  __typename?: 'Rating'
  rater: Scalars['String']
  value: Scalars['Int']
}
// ...
export type Topic = {
  __typename?: 'Topic'
  body?: Maybe<Scalars['String']>
  children?: Maybe<Array<Maybe<Scalars['String']>>>
  community: Scalars['String']
  parents?: Maybe<Array<Maybe<Scalars['String']>>>
  pic?: Maybe<Scalars['String']>
  slug: Scalars['String']
  title?: Maybe<Scalars['String']>
  topicStat?: Maybe<TopicStat>
}

export type TopicInput = {
  body?: InputMaybe<Scalars['String']>
  children?: InputMaybe<Array<InputMaybe<Scalars['String']>>>
  community: Scalars['String']
  pic?: InputMaybe<Scalars['String']>
  slug: Scalars['String']
  title?: InputMaybe<Scalars['String']>
}

export type TopicResult = {
  __typename?: 'TopicResult'
  error?: Maybe<Scalars['String']>
  topic?: Maybe<Topic>
}

export type TopicStat = {
  __typename?: 'TopicStat'
  authors: Scalars['Int']
  shouts: Scalars['Int']
  subscriptions: Scalars['Int']
  views: Scalars['Int']
}

export type User = {
  __typename?: 'User'
  bio?: Maybe<Scalars['String']>
  communities?: Maybe<Array<Maybe<Scalars['Int']>>>
  createdAt: Scalars['DateTime']
  email?: Maybe<Scalars['String']>
  emailConfirmed?: Maybe<Scalars['Boolean']>
  id: Scalars['Int']
  links?: Maybe<Array<Maybe<Scalars['String']>>>
  muted?: Maybe<Scalars['Boolean']>
  name?: Maybe<Scalars['String']>
  notifications?: Maybe<Array<Maybe<Scalars['Int']>>>
  oauth?: Maybe<Scalars['String']>
  old_id?: Maybe<Scalars['String']>
  password?: Maybe<Scalars['String']>
  ratings?: Maybe<Array<Maybe<Rating>>>
  slug: Scalars['String']
  updatedAt?: Maybe<Scalars['DateTime']>
  username: Scalars['String']
  userpic?: Maybe<Scalars['String']>
  wasOnlineAt?: Maybe<Scalars['DateTime']>
}